PR2487240 - Concurrent CP join race condition
This issue has been reported to upstream kubeadm: https://github.com/kubernetes/kubeadm/issues/2005
And this patch implements the suggested workaround.

As mentioned in PR2487240, this issue happens very rarely in a very small gap where the 3rd member
is performing etcd join while the 2nd member had joined but not yet fully started.

The symptom is that only the 3rd member's etcd container is not coming up and complaining

member count is unequal

The root cause is that when kubeadm is performing etcd member join, it calls etcd client's
MemberAdd function with the local peerURL and this function returns a list of etcd Member.

kubeadm expects that there should be 1 and ONLY 1 Member in this list that does not have Name field,
which is the currently-being-added member. And then it will insert the provided local name to the
Name field of those Members that don't have Name.

However, this issue happens because somehow there are more than 1 Member that do not have Name.
In this case, kubeadm code incorrectly inserted local name to all those Members that don't have Name.
Therefore, the resulting etcd manifest contains the value of --initial-cluster with duplicate Names,
in a 3-node etcd cluster, the incorrect etcd manifest will have 2 unique Name and 3 unique peerURLs.

And when the local etcd starts, it looks at the --initial-cluster and thinks that there are only 2
members, due to the duplicate Names. But actually there are 3 members, hence the member count is
unequal issue.

The fix here is to assign the etcd ID to Name if the member has no Name and peerURL does not match.