Port filtered certs to dynamic certs logic
This takes the filtered certificates patch that was applied to 1.15 and
1.16 and ports it to 1.174.

This is a comparitively major change to the patch as 1.16->1.16 introduced
dynamic certificates infrastructure to reload certs without restarting
apiserver.

I believe the dynamic certificate infrastructure wasn't written with complex
certificate serving methods, such as SNI or this patch in mind. I note this
because this patch follows the model demonstrated by the current code, even
when it adds rather convoluted paths.

The only code that has not significantly altered since the prior versions:
* cli flag parsing
* tls.Config.GetCertificate callback
* connection match logic